+++ packages/angular_devkit/build_angular/src/dev-server/index.ts
@@ -218,10 +218,7 @@ export function serveWebpackBrowser(
     }
 
     let locale: string | undefined;
-    if (browserOptions.i18nLocale) {
-      // Deprecated VE option
-      locale = browserOptions.i18nLocale;
-    } else if (i18n.shouldInline) {
+    if (i18n.shouldInline) {
       // Dev-server only supports one locale
       locale = [...i18n.inlineLocales][0];
     } else if (i18n.hasDefinedSourceLocale) {
