+++ packages/angular_devkit/build_angular/src/extract-i18n/index.ts
@@ -167,9 +167,6 @@ export async function execute(
         vendor: true,
       },
       buildOptimizer: false,
-      i18nLocale: i18n.sourceLocale,
-      i18nFormat: format,
-      i18nFile: outFile,
       aot: true,
       progress: options.progress,
       assets: [],
