+++ packages/angular_devkit/build_angular/src/utils/output-paths.ts
@@ -10,14 +10,9 @@ import { join } from 'path';
 import { I18nOptions } from './i18n-options';
 
 export function ensureOutputPaths(baseOutputPath: string, i18n: I18nOptions): Map<string, string> {
-  const outputPaths: [string, string][] =
-    i18n.shouldInline
-      ? [...i18n.inlineLocales].map(l => [l, i18n.flatOutput ? baseOutputPath : join(baseOutputPath, l)])
-      : [
-          i18n.veCompatLocale
-            ? [i18n.veCompatLocale, join(baseOutputPath, i18n.veCompatLocale)]
-            : ['', baseOutputPath],
-        ];
+  const outputPaths: [string, string][] = i18n.shouldInline
+    ? [...i18n.inlineLocales].map(l => [l, i18n.flatOutput ? baseOutputPath : join(baseOutputPath, l)])
+    : [['', baseOutputPath]];
 
   for (const [, outputPath] of outputPaths) {
     if (!existsSync(outputPath)) {
