+++ tests/legacy-cli/e2e/tests/i18n/ivy-localize-basehref.ts
@@ -8,7 +8,7 @@
 import { expectFileToMatch } from '../../utils/fs';
 import { ng } from '../../utils/process';
 import { updateJsonFile } from '../../utils/project';
-import { externalServer, langTranslations, setupI18nConfig } from './legacy';
+import { externalServer, langTranslations, setupI18nConfig } from './setup';
 
 const baseHrefs = {
   'en-US': '/en/',
@@ -18,7 +18,7 @@ const baseHrefs = {
 
 export default async function() {
   // Setup i18n tests and config.
-  await setupI18nConfig(true);
+  await setupI18nConfig();
 
   // Update angular.json
   await updateJsonFile('angular.json', workspaceJson => {
