+++ tests/legacy-cli/e2e/tests/i18n/ivy-localize-dl-xliff1.ts
@@ -1,9 +1,9 @@
 import { executeTest } from './ivy-localize-dl-xliff2';
-import { setupI18nConfig } from './legacy';
+import { setupI18nConfig } from './setup';
 
 export default async function() {
   // Setup i18n tests and config.
-  await setupI18nConfig(true, 'xlf');
+  await setupI18nConfig('xlf');
 
   // Execute the tests
   await executeTest();
