+++ tests/legacy-cli/e2e/tests/i18n/ivy-localize-es5.ts
@@ -2,7 +2,7 @@ import { expectFileNotToExist, expectFileToMatch, readFile } from '../../utils/f
 import { ng } from '../../utils/process';
 import { updateJsonFile } from '../../utils/project';
 import { expectToFail } from '../../utils/utils';
-import { externalServer, langTranslations, setupI18nConfig } from './legacy';
+import { externalServer, langTranslations, setupI18nConfig } from './setup';
 
 export default async function() {
   // Setup i18n tests and config.
