+++ tests/legacy-cli/e2e/tests/i18n/ivy-localize-locale-data.ts
@@ -7,11 +7,11 @@
  */
 import { ng } from '../../utils/process';
 import { updateJsonFile } from '../../utils/project';
-import { setupI18nConfig } from './legacy';
+import { setupI18nConfig } from './setup';
 
 export default async function() {
   // Setup i18n tests and config.
-  await setupI18nConfig(true);
+  await setupI18nConfig();
 
   // Update angular.json
   await updateJsonFile('angular.json', workspaceJson => {
