+++ tests/legacy-cli/e2e/tests/i18n/ivy-localize-sourcelocale.ts
@@ -8,11 +8,11 @@
 import { expectFileToMatch } from '../../utils/fs';
 import { ng } from '../../utils/process';
 import { updateJsonFile } from '../../utils/project';
-import { langTranslations, setupI18nConfig } from './legacy';
+import { langTranslations, setupI18nConfig } from './setup';
 
 export default async function() {
   // Setup i18n tests and config.
-  await setupI18nConfig(true);
+  await setupI18nConfig();
 
   // Update angular.json
   await updateJsonFile('angular.json', workspaceJson => {
